--- a/src/bluetooth_data_tools/gap.py
+++ b/src/bluetooth_data_tools/gap.py
@@ -208,19 +208,18 @@ def _uncached_parse_advertisement_data(
     tx_power: int | None = None
 
     for gap_data in data:
-        gap_view = gap_data
         offset = 0
         total_length = len(gap_data)
         while offset < total_length:
             try:
-                length = gap_view[offset]
+                length = gap_data[offset]
                 if not length:
                     if offset + 2 < total_length:
                         # Maybe zero padding
                         offset += 1
                         continue
                     break
-                gap_type_num = gap_view[offset + 1]
+                gap_type_num = gap_data[offset + 1]
                 if not gap_type_num:
                     break
                 start = offset + 2