instead of using a view, use char

before
```
Parsing 100000 bluetooth messages took 0.032907540997257456 seconds
Parsing 100000 bluetooth messages without caching took 1.9759795420104638 seconds
```


after
```
Parsing 100000 bluetooth messages took 0.024532374984119087 seconds
Parsing 100000 bluetooth messages without caching took 1.4332365000154823 seconds
```